Add files via upload
This is an E-Project that I received from Aptech. This is a website about COVID-19 provided by Aptech. It is basically a hospital website where all kinds of safety measures have been explained. There are special doctors who are qualified to treat this disease effectively. If we look at the databases provided, it can be seen that a complete dynamic website has been created.